% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coRPysprofiling.R
\name{corpora_best_match}
\alias{corpora_best_match}
\title{Returns a tibble of distances from the reference document for each corpus in a vector of corpora.
This tibble is sorted in the order of increasing distance.}
\usage{
corpora_best_match(
  refDoc,
  corpora,
  metric = "cosine_similarity",
  model_name = "cb_ns_500_10"
)
}
\arguments{
\item{refDoc}{character vector for reference document}

\item{corpora}{character vector for corpora}

\item{metric}{character vector for metric used to calculate distance, optional (default : "cosine_similarity")}

\item{model_name}{character vector, optional (default : "cb_ns_500_10")}
}
\value{
tibble
}
\description{
Returns a tibble of distances from the reference document for each corpus in a vector of corpora.
This tibble is sorted in the order of increasing distance.
}
\examples{
coRPysprofiling::corpora_best_match("kitten meows", c("ice cream is yummy", "cat meowed", "dog barks", "The Hitchhiker's Guide to the Galaxy has become an international multi-media phenomenon"))
}
