% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coRPysprofiling.R
\name{corpus_analysis}
\alias{corpus_analysis}
\title{Generate basic statistic for words from the input corpus}
\usage{
corpus_analysis(corpus)
}
\arguments{
\item{corpus}{character vector representing a corpus}
}
\value{
data.frame
}
\description{
Generate basic statistic for words from the input corpus
}
\examples{
coRPysprofiling::corpus_analysis("How many animals in Russia? and how many in US?")
coRPysprofiling::corpus_analysis("How many animals in Russia? and how many in US?")["word_total", ]
}
