% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coRPysprofiling.R
\name{corpus_viz}
\alias{corpus_viz}
\title{Generate visualizations for words from the input corpus}
\usage{
corpus_viz(corpus)
}
\arguments{
\item{corpus}{a character vector representing a corpus}
}
\value{
a list of a word cloud, a histogram of word length frequencies, and a histogram of word frequencies
}
\description{
Generate visualizations for words from the input corpus
}
\examples{
coRPysprofiling::corpus_viz("some text")
coRPysprofiling::corpus_viz("some text")['word cloud']
coRPysprofiling::corpus_viz("some text")['word freq bar chart']
coRPysprofiling::corpus_viz("some text")['word length bar chart']
}
