% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coRPysprofiling.R
\name{load_pretrained}
\alias{load_pretrained}
\title{Download and load pretrained word2vector models (https://github.com/maxoodf/word2vec#basic-usage)}
\usage{
load_pretrained(dir = "data", model_name = "cb_ns_500_10")
}
\arguments{
\item{dir}{character vector for name of dir where pretrained models will be downloaded, optional (default: "data")}

\item{model_name}{character vector for name of pretrained model, optional (default: "cb_ns_500_10")}
}
\value{
word2vec model object
}
\description{
Download and load pretrained word2vector models (https://github.com/maxoodf/word2vec#basic-usage)
}
\examples{
coRPysprofiling::load_pretrained()
coRPysprofiling::load_pretrained(model_name = "sg_hs_500_10")
}
