% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coRPysprofiling.R
\name{corpora_best_match}
\alias{corpora_best_match}
\title{Returns a tibble of distances from the reference document for each corpus in a vector of corpora.
This tibble is sorted in the order of increasing distance.}
\usage{
corpora_best_match(refDoc, corpora, metric = "cosine_similarity")
}
\arguments{
\item{refDoc}{character vector}

\item{corpora}{character vector}

\item{metric}{character vector, optional (default : "cosine_similarity")}
}
\value{
tibble
}
\description{
Returns a tibble of distances from the reference document for each corpus in a vector of corpora.
This tibble is sorted in the order of increasing distance.
}
\examples{
corpora_best_match(("kitten meows", c("ice cream is yummy", "cat meowed", "dog barks", "The Hitchhiker's Guide to the Galaxy has become an international multi-media phenomenon"))
}
