% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coRPysprofiling.R
\name{corpora_compare}
\alias{corpora_compare}
\title{Returns a numeric vector of the distance between the two corpora}
\usage{
corpora_compare(corpus1, corpus2, metric = "cosine_similarity")
}
\arguments{
\item{corpus1}{character vector}

\item{corpus2}{character vector}

\item{metric}{character vector, optional (default : "cosine_similarity")}
}
\value{
numeric vector
}
\description{
Returns a numeric vector of the distance between the two corpora
}
\examples{
corpora_compare("kitten meows", "ice cream is yummy")
}
