% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coRPysprofiling.R
\name{corpus_analysis}
\alias{corpus_analysis}
\title{Generate basic statistic for words from the input corpus}
\usage{
corpus_analysis(corpus)
}
\arguments{
\item{corpus}{a string representing a corpus}
}
\value{
tibble
}
\description{
Generate basic statistic for words from the input corpus
}
\examples{
corpus_analysis("How many animals in Russia?")
}
