% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coRPysprofiling.R
\name{corpus_viz}
\alias{corpus_viz}
\title{Generate visualizations for words from the input corpus}
\usage{
corpus_viz(corpus, display = TRUE)
}
\arguments{
\item{corpus}{a string representing a corpus}

\item{display}{a boolean (optional); If display is False, the plots will be hidden from the output}
}
\value{
a dictionary of a word cloud, and a data frame, which can be used to draw a bar chart for words and word lengths
}
\description{
Generate visualizations for words from the input corpus
}
\examples{
coRPysprofiling::corpus_viz("some text")
}
